Added Supply Capped metric.
We are now able to estimate the amount of time a player
is unable to produce new units because he/she has not
built enough supply depots or command centers.  This should
have a small negative correlation with winning chances.

Note that for the future it may be handy to ignore the first
1-2% of the match as it is not as complex a matter to become
supply capped during that time.  Still, the precision gain would
be tiny...